https://github.com/bitfaster/BitFaster.Caching/pull/378 introduced a bug by moving items from hot instead of cold.

This caused WhenItemsAreScannedInParallelCapacityIsNotExceeded to fail intermittently:

![image](https://github.com/bitfaster/BitFaster.Caching/assets/12851828/58321257-4321-4c18-898c-02cef9da1472)