@bitfaster 2023-08-18
Re-run hit rate suite on this PR. Before numbers come from 2.2.0 (not https://github.com/bitfaster/BitFaster.Caching/pull/378)

## Wikibench (worse)

| CacheSize | Before               | After                |
|-----------|--------------------  |----------------------|
|        25 |                30.75 |   28.980103382537575 |
|        50 |                38.54 |   37.299943707435645 |
|        75 |                42.52 |   41.427949141134235 |
|       100 |                45.1  |    44.06817855702613 |
|       125 |                47.22 |     46.1382063286935 |
|       150 |                48.69 |    47.62653088961094 |
|       175 |                49.93 |    48.87486783957522 |
|       200 |                50.83 |     49.8392244670742 |

## Glimpse (unchanged)

| CacheSize | Before | After |
|-----------|--------------------|----------------------|
|       250 |              12.83 |   12.834580216126351 |
|       500 |              26.13 |   26.134663341645886 |
|       750 |              39.43 |    39.43474646716542 |
|      1000 |              43.28 |   43.275145469659186 |
|      1250 |              49.93 |   49.925187032418954 |
|      1500 |              52.4  |    52.40232751454697 |
|      1750 |              53.5  |    53.49958437240233 |
|      2000 |              53.55 |    53.54945968412302 |