@bitfaster 2023-08-22
Cycle benchmark:

Before

|             Method |            Runtime |     Mean | Ratio | Allocated |
|------------------- |------------------- |---------:|------:|----------:|
|  FastConcurrentLru |           .NET 6.0 | 23.41 us |  1.00 |   8.51 KB |
|      ConcurrentLru |           .NET 6.0 | 25.55 us |  1.09 |   8.51 KB |
| ConcurrentLruEvent |           .NET 6.0 | 25.91 us |  1.11 |  12.29 KB |
| FastConcurrentTLru |           .NET 6.0 | 24.22 us |  1.03 |   9.45 KB |
|     ConcurrentTLru |           .NET 6.0 | 26.28 us |  1.12 |   9.45 KB |
|         ClassicLru |           .NET 6.0 | 16.97 us |  0.72 |     14 KB |
|                    |                    |          |       |           |
|  FastConcurrentLru | .NET Framework 4.8 | 37.24 us |  1.00 |  14.34 KB |
|      ConcurrentLru | .NET Framework 4.8 | 38.63 us |  1.04 |  14.34 KB |
| ConcurrentLruEvent | .NET Framework 4.8 | 39.98 us |  1.07 |  18.13 KB |
| FastConcurrentTLru | .NET Framework 4.8 | 48.89 us |  1.31 |  15.29 KB |
|     ConcurrentTLru | .NET Framework 4.8 | 51.07 us |  1.37 |  15.29 KB |
|         ClassicLru | .NET Framework 4.8 | 22.17 us |  0.60 |  17.05 KB |

After
|             Method |            Runtime |     Mean | Ratio | Allocated |
|------------------- |------------------- |---------:|------:|----------:|
|  FastConcurrentLru |           .NET 6.0 | 23.84 us |  1.00 |   8.51 KB |
|      ConcurrentLru |           .NET 6.0 | 25.72 us |  1.08 |   8.51 KB |
| ConcurrentLruEvent |           .NET 6.0 | 25.58 us |  1.07 |  12.29 KB |
| FastConcurrentTLru |           .NET 6.0 | 24.13 us |  1.01 |   9.45 KB |
|     ConcurrentTLru |           .NET 6.0 | 26.20 us |  1.10 |   9.45 KB |
|         ClassicLru |           .NET 6.0 | 17.34 us |  0.73 |     14 KB |
|                    |                    |          |       |           |
|  FastConcurrentLru | .NET Framework 4.8 | 37.73 us |  1.00 |  14.34 KB |
|      ConcurrentLru | .NET Framework 4.8 | 39.17 us |  1.04 |  14.34 KB |
| ConcurrentLruEvent | .NET Framework 4.8 | 39.94 us |  1.06 |  18.13 KB |
| FastConcurrentTLru | .NET Framework 4.8 | 48.25 us |  1.28 |  15.29 KB |
|     ConcurrentTLru | .NET Framework 4.8 | 50.89 us |  1.35 |  15.29 KB |
|         ClassicLru | .NET Framework 4.8 | 21.84 us |  0.58 |  17.05 KB |